% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayons-package.R
\docType{package}
\name{crayons-package}
\alias{crayons-package}
\title{crayons: Color Palettes from Crayon Boxes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides color palettes based on crayon colors since the early 1900s. Colors are based on various crayon colors, sets, and promotional palettes, most of which can be found at \url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}. All palettes are discrete palettes and are not necessarily color-blind friendly. Provides scales for 'ggplot2' for discrete coloring.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/crayons}
  \item \url{http://christophertkenny.com/crayons/}
  \item Report bugs at \url{https://github.com/christopherkenny/crayons/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
