% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonbox.R
\name{crayons}
\alias{crayons}
\title{crayons}
\usage{
crayons()
}
\value{
Nothing. It's side effect is to activate the Viewer to be used as a coloring tool.
}
\description{
During an interaction session, create a coloring tool in the Viewer for
users to create different vectors of colors. Will copy directly to the user clipboard.
}
