% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonbox.R
\name{get_plots}
\alias{get_plots}
\title{get_plots}
\usage{
get_plots()
}
\value{
A list of plot file names as full path.
}
\description{
Will get the current plots that have been made in
the interactive R session -- helps facilitate what plots need to be customized.
}
