% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_colors.R
\name{pick_colors}
\alias{pick_colors}
\title{Pick \code{n} Colors}
\usage{
pick_colors(hex, n)
}
\arguments{
\item{hex}{a character vector of hexademical colors}

\item{n}{the integer number of desired colors}
}
\value{
a character vector of length n
}
\description{
Places each color into RGB space, clusters them with kmeans, and returns the
colors nearest each cluster center.
}
\examples{
pick_colors(crayons$standard, 10)
}
