% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colors}
\usage{
plot_palette(x)
}
\arguments{
\item{x}{A vector of colors.}
}
\value{
A \code{ggplot}
}
\description{
Uses \code{ggplot2} for a non-base-plot implementation of \code{scales::show_col()}.
Plots 6 digit hex code over colors.
}
\examples{
plot_palette(crayons$standard16)
}
