% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_crayons}
\alias{scale_color_crayons}
\alias{scale_fill_crayons}
\alias{scale_colour_crayons}
\title{Crayon Color Scales for \code{ggplot2}}
\usage{
scale_color_crayons(palette = "standard16", which = NULL, ..., reverse = FALSE)

scale_fill_crayons(palette = "standard16", which = NULL, ..., reverse = FALSE)

scale_colour_crayons(
  palette = "standard16",
  which = NULL,
  ...,
  reverse = FALSE
)
}
\arguments{
\item{palette}{TODO}

\item{which}{TODO}

\item{...}{arguments passed on to [ggplot2::discrete_scale()}

\item{reverse}{TODO}
}
\description{
Crayon Color Scales for \code{ggplot2}
}
\examples{
library(ggplot2)
#TODO

}
